chore: Fix feature flags and build docs
