+++ tonic-build/src/client.rs
@@ -3,7 +3,10 @@ use crate::{generate_doc_comments, naive_snake_case};
 use proc_macro2::TokenStream;
 use quote::{format_ident, quote};
 
-/// Generate service for client
+/// Generate service for client.
+///
+/// This takes some `Service` and will generate a `TokenStream` that contains
+/// a public module with the generated client.
 pub fn generate<T: Service>(service: &T, proto_path: &str) -> TokenStream {
     let service_ident = quote::format_ident!("{}Client", service.name());
     let client_mod = quote::format_ident!("{}_client", naive_snake_case(&service.name()));
