+++ tonic-build/src/server.rs
@@ -4,7 +4,10 @@ use proc_macro2::{Span, TokenStream};
 use quote::quote;
 use syn::{Ident, Lit, LitStr};
 
-/// Generate service for Server
+/// Generate service for Server.
+///
+/// This takes some `Service` and will generate a `TokenStream` that contains
+/// a public module containing the server service and handler trait.
 pub fn generate<T: Service>(service: &T, proto_path: &str) -> TokenStream {
     let methods = generate_methods(service, proto_path);
