+++ tonic/src/codec/mod.rs
@@ -6,18 +6,15 @@
 mod buffer;
 mod decode;
 mod encode;
-#[cfg(feature = "data-prost")]
+#[cfg(feature = "prost")]
 mod prost;
-
-#[cfg(all(test, feature = "data-prost"))]
-mod prost_tests;
 
 use std::io;
 
 pub use self::decode::Streaming;
 pub(crate) use self::encode::{encode_client, encode_server};
-#[cfg(feature = "data-prost")]
-#[cfg_attr(docsrs, doc(cfg(feature = "data-prost")))]
+#[cfg(feature = "prost")]
+#[cfg_attr(docsrs, doc(cfg(feature = "prost")))]
 pub use self::prost::ProstCodec;
 use crate::Status;
 pub use buffer::{DecodeBuf, EncodeBuf};
